fix(admin-extension-sdk): publish the extension slot vocabulary from one place
The list of admin UI extension slots was maintained twice: in
packages/admin/src/ui-extensions/registry.ts, for the shell that renders them,
and in packages/apps/src/contracts.ts, as the enum the app manifest schema
validates against. Both held the same eight identifiers and agreed only by
discipline. A slot added to one would have been rejected by the schema or left
unrendered by the shell, and nothing checked.

Move the list into @voyant-travel/admin-extension-sdk and derive both from it.
That package is the right owner: it is dependency-free, it is what an extension
author installs, and the slot vocabulary is part of the public extension
contract rather than an internal detail of the shell. Adding a slot is now one
edit, versioned with the contract it widens.

Both packages keep their existing export names, so no consumer import changes.
