+++ packages/admin/src/ui-extensions/registry.ts
@@ -5,22 +5,25 @@
  * UI extension is contributed through the same {@link AdminWidgetSlot}
  * mechanism as any first-party widget. Adding a slot is a MINOR change (a new
  * capability extensions can opt into); renaming or removing one is MAJOR.
+ *
+ * The list itself lives in `@voyant-travel/admin-extension-sdk`, which is the
+ * versioned contract an extension author installs. This module adapts it to
+ * the shell's widget vocabulary.
  */
+import {
+  type AdminUiExtensionSlot,
+  ADMIN_UI_EXTENSION_SLOTS as CONTRACT_SLOTS,
+} from "@voyant-travel/admin-extension-sdk/types"
 import type { AdminWidgetSlot } from "../extensions.js"
 
-export const ADMIN_UI_EXTENSION_SLOTS = [
-  "dashboard.header",
-  "dashboard.after-kpis",
-  "dashboard.footer",
-  "booking.details.header",
-  "booking.details.after-summary",
-  "invoice.details.header",
-  "invoice.details.after-summary",
-  "workspace.header.actions",
-] as const satisfies readonly AdminWidgetSlot[]
+/**
+ * Re-exported from the contract package so the shell and the manifest schema
+ * cannot drift apart. The `satisfies` keeps the host's own guarantee: every
+ * contract slot must still be a usable {@link AdminWidgetSlot}.
+ */
+export const ADMIN_UI_EXTENSION_SLOTS = CONTRACT_SLOTS satisfies readonly AdminWidgetSlot[]
 
-/** A slot id from the public UI-extension registry. */
-export type AdminUiExtensionSlot = (typeof ADMIN_UI_EXTENSION_SLOTS)[number]
+export type { AdminUiExtensionSlot }
 
 /** Whether `slot` is one of the public UI-extension slots. */
 export function isAdminUiExtensionSlot(slot: string): slot is AdminUiExtensionSlot {
